+++ test/SaveHandler/MongoDBOptionsTest.php
@@ -1,19 +1,17 @@
 <?php
 
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-session for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-session/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-session/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Session\SaveHandler;
+namespace LaminasTest\Session\SaveHandler;
 
-use Zend\Session\SaveHandler\MongoDBOptions;
+use Laminas\Session\SaveHandler\MongoDBOptions;
 
 /**
- * @group      Zend_Session
+ * @group      Laminas_Session
  */
 class MongoDBOptionsTest extends \PHPUnit_Framework_TestCase
 {
@@ -73,7 +71,7 @@ class MongoDBOptionsTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Zend\Session\Exception\InvalidArgumentException
+     * @expectedException Laminas\Session\Exception\InvalidArgumentException
      */
     public function testInvalidDatabase()
     {
@@ -83,7 +81,7 @@ class MongoDBOptionsTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Zend\Session\Exception\InvalidArgumentException
+     * @expectedException Laminas\Session\Exception\InvalidArgumentException
      */
     public function testInvalidCollection()
     {
@@ -93,7 +91,7 @@ class MongoDBOptionsTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Zend\Session\Exception\InvalidArgumentException
+     * @expectedException Laminas\Session\Exception\InvalidArgumentException
      */
     public function testInvalidSaveOptions()
     {
@@ -103,7 +101,7 @@ class MongoDBOptionsTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Zend\Session\Exception\InvalidArgumentException
+     * @expectedException Laminas\Session\Exception\InvalidArgumentException
      */
     public function testInvalidNameField()
     {
@@ -113,7 +111,7 @@ class MongoDBOptionsTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Zend\Session\Exception\InvalidArgumentException
+     * @expectedException Laminas\Session\Exception\InvalidArgumentException
      */
     public function testInvalidModifiedField()
     {
@@ -123,7 +121,7 @@ class MongoDBOptionsTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Zend\Session\Exception\InvalidArgumentException
+     * @expectedException Laminas\Session\Exception\InvalidArgumentException
      */
     public function testInvalidLifetimeField()
     {
@@ -133,7 +131,7 @@ class MongoDBOptionsTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Zend\Session\Exception\InvalidArgumentException
+     * @expectedException Laminas\Session\Exception\InvalidArgumentException
      */
     public function testInvalidDataField()
     {